+++ app/Http/Controllers/TagController.php
@@ -4,9 +4,19 @@ namespace App\Http\Controllers;
 
 use App\Models\Tag;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class TagController extends Controller
 {
+
+    public function __construct(){
+        $this->middleware('auth')->except(['index']);
+
+        // Sử dụng AdminMiddleware
+        $this->middleware('admin')->except(['index']);
+    }
+
+
     /**
      * Display a listing of the resource.
      *
@@ -21,7 +31,7 @@ class TagController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Http\Response
      */
     public function create()
     {
@@ -70,11 +80,14 @@ class TagController extends Controller
      * Show the form for editing the specified resource.
      *
      * @param  \App\Models\Tag  $tag
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Http\Response
      */
     public function edit(Tag $tag)
     {
-        //
+        // 87
+        abort_unless(Gate::allows('update', $tag), 403);
+
+
         return view('tag.edit')->with([
             'tag'=>$tag
         ]);
@@ -89,7 +102,9 @@ class TagController extends Controller
      */
     public function update(Request $request, Tag $tag)
     {
-        //
+        // 87
+        abort_unless(Gate::allows('update', $tag), 403);
+
         $request->validate([
             'name' => 'required',
             'style' => 'required',
@@ -113,6 +128,10 @@ class TagController extends Controller
      */
     public function destroy(Tag $tag)
     {
+
+        // 87
+        abort_unless(Gate::allows('delete', $tag), 403);
+
         // Lấy tên tag trước khi xóa
         $oldName = $tag->name;
         $tag->delete();
