+++ app/Http/Controllers/hobbyTagController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Tag;
 use App\Models\Hobby;
+use Illuminate\Support\Facades\Gate;
 
 class hobbyTagController extends Controller
 {
@@ -35,6 +36,11 @@ class hobbyTagController extends Controller
     public function attachTag($hobby_id, $tag_id){
         $hobby = Hobby::find($hobby_id);
 
+        // 89. Trước khi lấy ra tag cần gắn thì cần kiểm tra (với Gate)
+        if(Gate::denies('connect_hobbyTag', $hobby)){
+            abort(403, 'Oh no! This hobby is not yours!');
+        }
+
         // Lấy tag ra để thông báo gắn tag nào thành công
         $tag = Tag::find($tag_id);
 
@@ -49,6 +55,12 @@ class hobbyTagController extends Controller
 
     public function detachTag($hobby_id, $tag_id){
         $hobby = Hobby::find($hobby_id);
+
+        // 89
+        if(Gate::denies('connect_hobbyTag', $hobby)){
+            abort(403, 'Oh no! This hobby is not yours!');
+        }
+
         $tag = Tag::find($tag_id);
 
         $hobby->tags()->detach($tag_id);
